Add PXNavigationController
