+++ MercadoPagoSDK/MercadoPagoSDK/MPXLoadingViewController.swift
@@ -12,6 +12,5 @@ class MPXLoadingViewController: MercadoPagoUIViewController {
 
     open override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        MercadoPagoCheckout.firstViewControllerPushed = false
     }
 }
