+++ MercadoPagoSDK/MercadoPagoSDK/MercadoPagoUIViewController.swift
@@ -54,7 +54,6 @@ open class MercadoPagoUIViewController: UIViewController, UIGestureRecognizerDel
         navigationController?.interactivePopGestureRecognizer?.isEnabled = false
 
         self.loadMPStyles()
-        MercadoPagoCheckout.firstViewControllerPushed = true
 
         if shouldHideNavigationBar {
             navigationController?.setNavigationBarHidden(true, animated: false)
