+++ MercadoPagoSDK/MercadoPagoSDK/PXNavigationController.swift
@@ -1,0 +1,144 @@
+//
+//  PXNavigationController.swift
+//  MercadoPagoSDK
+//
+//  Created by Eden Torres on 03/05/2018.
+//  Copyright © 2018 MercadoPago. All rights reserved.
+//
+
+import Foundation
+import Foundation
+
+class PXNavigationController: NSObject {
+
+    var countLoadings: Int = 0
+    var navigationController: UINavigationController!
+    var viewControllerBase: UIViewController?
+    private var currentLoadingView: UIViewController?
+    private var rootViewController: UIViewController?
+
+    init(navigationController: UINavigationController) {
+        self.navigationController = navigationController
+        if self.navigationController.viewControllers.count > 0 {
+            let  newNavigationStack = self.navigationController.viewControllers.filter {!$0.isKind(of: MercadoPagoUIViewController.self) || $0.isKind(of: PXReviewViewController.self)
+            }
+            viewControllerBase = newNavigationStack.last
+        }
+    }
+
+    public func goToRootViewController() {
+        if let rootViewController = viewControllerBase {
+            self.navigationController.popToViewController(rootViewController, animated: true)
+            self.navigationController.setNavigationBarHidden(false, animated: false)
+        } else {
+            self.navigationController.dismiss(animated: true, completion: {
+                self.navigationController.setNavigationBarHidden(false, animated: false)
+            })
+        }
+    }
+
+    func presentLoading(animated: Bool = false, completion: (() -> Swift.Void)? = nil) {
+        self.countLoadings += 1
+        if self.countLoadings == 1 {
+            let when = DispatchTime.now() //+ 0.3
+            DispatchQueue.main.asyncAfter(deadline: when) {
+                if self.countLoadings > 0 && self.currentLoadingView == nil {
+                    self.createCurrentLoading()
+                    self.currentLoadingView?.modalTransitionStyle = .crossDissolve
+                    self.navigationController.present(self.currentLoadingView!, animated: true, completion: completion)
+                }
+            }
+        }
+    }
+
+    func presentInitLoading() {
+        self.createCurrentLoading()
+        self.currentLoadingView?.modalTransitionStyle = .crossDissolve
+        self.navigationController.present(self.currentLoadingView!, animated: false, completion: nil)
+    }
+
+    func dismissLoading(animated: Bool = true, finishCallback:(()-> Void)? = nil) {
+        self.countLoadings = 0
+        if self.currentLoadingView != nil {
+            self.currentLoadingView?.modalTransitionStyle = .crossDissolve
+            self.currentLoadingView!.dismiss(animated: animated, completion: {
+                self.currentLoadingView = nil
+                if let callback = finishCallback {
+                    callback()
+                }
+            })
+        }
+    }
+
+    internal func createCurrentLoading() {
+        let vcLoading = MercadoPagoUIViewController()
+        vcLoading.view.backgroundColor = ThemeManager.shared.getTheme().loadingComponent().backgroundColor
+        let loadingInstance = LoadingOverlay.shared.showOverlay(vcLoading.view, backgroundColor: ThemeManager.shared.getTheme().loadingComponent().backgroundColor, indicatorColor: ThemeManager.shared.getTheme().loadingComponent().tintColor)
+        vcLoading.view.addSubview(loadingInstance)
+        loadingInstance.bringSubview(toFront: vcLoading.view)
+        self.currentLoadingView = vcLoading
+    }
+
+    internal func pushViewController(viewController: MercadoPagoUIViewController,
+                                     animated: Bool, backToChechoutRoot: Bool = false) {
+
+        viewController.hidesBottomBarWhenPushed = true
+        // let mercadoPagoViewControllers = self.navigationController.viewControllers.filter {$0.isKind(of:MercadoPagoUIViewController.self)}
+        // Se remueve el comportamiento custom para el back. Ahora el back respeta el stack de navegacion, no hace popToX view controller
+        if backToChechoutRoot {
+            self.navigationController.navigationBar.isHidden = false
+            viewController.callbackCancel = { [weak self] in self?.backToCheckoutRoot() }
+        }
+
+        self.navigationController.pushViewController(viewController, animated: animated)
+        self.cleanCompletedCheckoutsFromNavigationStack()
+        self.dismissLoading()
+    }
+
+    func popViewController(animated: Bool = true) {
+        navigationController.popViewController(animated: animated)
+    }
+
+    func backToCheckoutRoot() {
+        let mercadoPagoViewControllers = self.navigationController.viewControllers.filter {$0.isKind(of: MercadoPagoUIViewController.self)}
+        if !mercadoPagoViewControllers.isEmpty {
+            self.navigationController.popToViewController(mercadoPagoViewControllers[0], animated: true)
+        }
+
+    }
+    internal func removeRootLoading() {
+        let currentViewControllers = self.navigationController.viewControllers.filter { (viewController: UIViewController) -> Bool in
+            return viewController != self.rootViewController
+        }
+        self.navigationController.viewControllers = currentViewControllers
+    }
+
+    public func popToWhenFinish(viewController: UIViewController) {
+        if self.navigationController.viewControllers.contains(viewController) {
+            self.viewControllerBase = viewController
+        }
+    }
+
+    func cleanCompletedCheckoutsFromNavigationStack() {
+        let  pxResultViewControllers = self.navigationController.viewControllers.filter {$0.isKind(of: PXResultViewController.self)}
+        if let lastResultViewController = pxResultViewControllers.last {
+            let index = self.navigationController.viewControllers.index(of: lastResultViewController)
+            let  validViewControllers = self.navigationController.viewControllers.filter {!$0.isKind(of: MercadoPagoUIViewController.self) || self.navigationController.viewControllers.index(of: $0)! > index! || $0 == self.navigationController.viewControllers.last }
+            self.navigationController.viewControllers = validViewControllers
+        }
+    }
+}
+
+extension PXNavigationController: UINavigationControllerDelegate {
+
+    func suscribeToNavigationFlow() {
+        navigationController.delegate = self
+    }
+
+    public func navigationController(_ navigationController: UINavigationController, willShow viewController: UIViewController, animated: Bool) {
+        if !(viewController is MercadoPagoUIViewController) {
+            ThemeManager.shared.applyAppNavBarStyle(navigationController: navigationController)
+            PXCheckoutStore.sharedInstance.clean()
+        }
+    }
+}
